Abstract detectUbuntuCodename
