+++ lib/native_loader.js
@@ -17,26 +17,9 @@
 const fs = require('fs');
 const path = require('path');
 const debug = require('debug')('rclnodejs');
+const { detectUbuntuCodename } = require('./utils');
 
 let nativeModule = null;
-
-/**
- * Detect Ubuntu codename from /etc/os-release
- * @returns {string|null} Ubuntu codename (e.g., 'noble', 'jammy') or null
- */
-function detectUbuntuCodename() {
-  if (process.platform !== 'linux') {
-    return null;
-  }
-
-  try {
-    const osRelease = fs.readFileSync('/etc/os-release', 'utf8');
-    const match = osRelease.match(/^VERSION_CODENAME=(.*)$/m);
-    return match ? match[1].trim() : null;
-  } catch {
-    return null;
-  }
-}
 
 // Simplified loader: only use prebuilt binaries with exact Ubuntu/ROS2/arch match
 // Note: Prebuilt binaries are only supported on Linux (Ubuntu) platform
